% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{fizzbuzz}
\alias{fizzbuzz}
\title{A function that takes in a vector of numbers are returns variations of the
string "Fizz" or "Buzz" depending on the value of the numbers.}
\usage{
fizzbuzz(nums)
}
\arguments{
\item{nums}{A vector of numbers to run against FizzBuzz function.}
}
\value{
A character vector returned that represents the FizzBuzz logic.
}
\description{
A function that takes in a vector of numbers are returns variations of the
string "Fizz" or "Buzz" depending on the value of the numbers.
}
