% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{A series of Flagstaff Pulliam Airport Weather Station Data from 1950 - 2019.}
\format{
a data frame with  observations
\describe{
  \item{DATE}{The date of observation as a POSIX date format.}
  \item{Rain}{Daily precipitation measures.}
  \item{SNOW}{Daily snow measures.}
  \item{TMIN}{Daily temperature low.}
  \item{TMAx}{Daily temperature high.}
}
}
\source{
\url{www.ncdc.noaa.gov}
}
\usage{
Flagstaff_Weather
}
\description{
A series of Flagstaff Pulliam Airport Weather Station Data from 1950 - 2019.
}
\keyword{datasets}
